Write API booking service

+++ src/migrations/20220228192926-create-import-details.js
@@ -18,7 +18,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       price: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.FLOAT
       },
       createdAt: {
         allowNull: false,
